cleanup and reuse the liquidity section on both market and pool page
